Move hero roster building function to store
