+++ src/state/roster/index.ts
@@ -69,7 +69,7 @@ const initialState = {
     num_units: 0,
     points: 0,
     bow_count: 0,
-    leader_warband_num: null,
+    leader_warband_id: null,
     warbands: [],
   },
   factions: [],
